added request bodies to use when dealing with registration and login requests.
