+++ cmd/app/request_bodies/auth_request_bodies.go
@@ -1,0 +1,18 @@
+package request_bodies
+
+/* Registration */
+
+type RegisterRequest struct {
+	Email string `json:"email"`
+	Username string `json:"username"`
+	Password string `json:"password"`
+	ConfirmPwd string `json:"confirm_pwd"`
+}
+
+/* Login */
+
+type LoginRequest struct {
+	Username string `json:"username"`
+	Password string `json:"password"`
+}
+
